Split four more oversized sync files by topic
NetSyncSystem/Realize.cs (1757), the occupancy Capture.cs (1302),
VisualCustomizationSyncSystem.cs (1146) and RouteSyncSystem/Realize.cs (1110)
each held several subjects at once. Split along the seams they already had, with
each file noting what belongs in it:

  net realize      -> operation assembly, unresolved holds, span geometry
  occupancy capture-> departures, the bucket scan, entity reads, hashing/trace
  visual custom.   -> capture, apply, appearance state
  route realize    -> connections, matching, commit

VisualCustomizationSyncSystem gets its own folder, as every other multi-file
system in the tree has.

One thing deliberately left alone: NetSyncSystem's RealizeIncoming is still an
819-line method, and the file is still the largest here because of it. It holds
about twenty-five interdependent locals across a try/finally that disposes
native collections, so breaking it up is a real refactor with a real chance of
changing behaviour - not something to do in the same pass as mechanical moves.

Pure code movement, verified member by member against the previous revision.
