+++ CS2MultiplayerMod/Game/Sync/Systems/Nets/NetSyncSystem/Realize.cs
@@ -20,6 +20,11 @@ namespace CS2MultiplayerMod.Game.Sync.Systems.Net
     // resolve captured native targets (or classify fallback geometry), then route every course
     // through one serialized Temp+ApplyTool transaction. Dependent systems wait for its drain so
     // they never observe half-realized network geometry.
+    //
+    // This file holds the operation state and RealizeIncoming, the cycle itself. Assembling an
+    // operation out of its messages is in RealizeOperation.cs, holding one whose targets have not
+    // arrived is in RealizeHold.cs, and the span and endpoint geometry it leans on is in
+    // RealizeSpan.cs.
     public partial class NetSyncSystem
     {
         private const long OperationAssemblyWindowMs = 3000;
@@ -962,796 +967,6 @@ namespace CS2MultiplayerMod.Game.Sync.Systems.Net
                 }
                 Diagnostics.FlightRecorder.Note("net build batch armed n=" + built + (splitUsed ? " +split" : ""));
             }
-        }
-
-        private void PruneCompletedNetOperations(long now)
-        {
-            _completedNetOperations.Prune(now);
-            _armedNetOperations.Prune(now);
-        }
-
-        /// <summary>
-        /// Re-queue <paramref name="work"/>[<paramref name="from"/>..] ahead of the shared inbox.
-        /// </summary>
-        private void RequeueFrom(List<SimulationCommandMessage> work, int from)
-        {
-            if (from < work.Count)
-                _remoteDeferred.InsertRange(0, work.GetRange(from, work.Count - from));
-        }
-
-        private static bool HasExternalNativeTarget(NetEndpointTargetKind kind) =>
-            kind == NetEndpointTargetKind.Node || kind == NetEndpointTargetKind.Edge ||
-            kind == NetEndpointTargetKind.OwnedNode || kind == NetEndpointTargetKind.OwnedEdge;
-
-        private static string DescribeUnresolvedEndpoint(NetPlacementCommand command,
-            bool startResolved)
-        {
-            NetEndpointIntent failed = startResolved ? command.End : command.Start;
-            return "course=" + command.CourseIndex + " " + (startResolved ? "end" : "start") +
-                   " kind=" + failed.Kind + " prefab='" + failed.TargetPrefabName + "' anchor=" +
-                   failed.AnchorX.ToString("F1") + "," + failed.AnchorY.ToString("F1") + "," +
-                   failed.AnchorZ.ToString("F1");
-        }
-
-        /// <summary>
-        /// Record that a resolved endpoint will split <paramref name="target"/>, and report whether
-        /// that claim is consistent with the source operation.
-        ///
-        /// Several courses of one operation may legitimately tap the SAME source edge: CourseSplitSystem
-        /// receives them together and cuts that edge once into all of its pieces. Two courses that named
-        /// DIFFERENT source edges but land on the same local edge are a different matter - this machine
-        /// never received the split that separated them. Committing both would hand the apply pass two
-        /// Temps sharing one original, which it dereferences without a liveness check.
-        /// </summary>
-        private bool TryClaimSplitTarget(NetEndpointIntent intent, Entity target, int kind)
-        {
-            if (kind != KindSplit || target == Entity.Null) return true;
-            Bezier4x3 source = TargetCurveOf(intent);
-            Bezier4x3 claimed;
-            if (!_batchSplitClaims.TryGetValue(target, out claimed))
-            {
-                _batchSplitClaims[target] = source;
-                return true;
-            }
-            return SameCurveBits(claimed, source) || SameCurveBitsReversed(claimed, source);
-        }
-
-        /// <summary>
-        /// Pull one complete source operation from the ordered command streams. Messages belonging
-        /// to later operations may be encountered while waiting for an interleaved course; they are
-        /// returned to the simulation-thread prefix in their original order. An incomplete operation
-        /// waits briefly and is then dropped as a whole, never realized as broken geometry.
-        /// </summary>
-        private bool TryTakeCompleteOperation(MultiplayerSession session, long now,
-            out List<SimulationCommandMessage> operation, out bool nativeOperation,
-            out NetToolOperationCommand mixedOperation)
-        {
-            operation = null;
-            nativeOperation = false;
-            mixedOperation = null;
-
-            const int MaxScan = NetInboxCap;
-            var scanned = new List<SimulationCommandMessage>();
-            NetOperationKey key = default(NetOperationKey);
-            int expected = 0;
-            SimulationCommandMessage[] courses = null;
-            NetPlacementCommand[] decodedCourses = null;
-            int received = 0;
-
-            for (int scan = 0; scan < MaxScan && (expected == 0 || received < expected); scan++)
-            {
-                SimulationCommandMessage message;
-                if (!TryTakeNextPlacementMessage(out message)) break;
-                if (message.OriginPlayerId == session.LocalPlayerId) continue;
-
-                if (message.CommandId == NetToolOperationCommand.Id)
-                {
-                    if (expected == 0)
-                    {
-                        try { mixedOperation = NetToolOperationCommand.Decode(message.Body); }
-                        catch (System.Exception ex)
-                        {
-                            Mod.log.Warn("[MP] NetSync: dropping malformed mixed net operation: " +
-                                         ex.Message);
-                            SyncInbox.RequestResync(Diagnostics.ResyncReport
-                                .Create("malformed mixed net operation", "net",
-                                    Diagnostics.ResyncEvidence.StreamLoss)
-                                .About("mixed operation from player " + message.OriginPlayerId)
-                                .Tried("nothing - the operation could not be decoded")
-                                .Fact("decoder said", ex.Message));
-                            return false;
-                        }
-                        operation = new List<SimulationCommandMessage>(1) { message };
-                        return true;
-                    }
-
-                    // It arrived after the first fragment of an older placement operation. Keep it
-                    // in the ordered prefix while scanning for that older operation's remaining
-                    // fragments; it will be the next operation realized, never overtaken.
-                    scanned.Add(message);
-                    continue;
-                }
-                if (message.CommandId != NetPlacementCommand.Id)
-                {
-                    Mod.log.Warn("[MP] NetSync: dropping unsupported queued command " +
-                                 message.CommandId + ".");
-                    continue;
-                }
-
-                NetPlacementCommand command;
-                try { command = NetPlacementCommand.Decode(message.Body); }
-                catch (System.Exception ex)
-                {
-                    Mod.log.Warn("[MP] NetSync: dropping malformed command: " + ex.Message);
-                    continue;
-                }
-
-                scanned.Add(message);
-                if (expected == 0)
-                {
-                    key = new NetOperationKey
-                    {
-                        Origin = message.OriginPlayerId,
-                        Operation = command.OperationId,
-                    };
-                    expected = command.CourseCount;
-                    courses = new SimulationCommandMessage[expected];
-                    decodedCourses = new NetPlacementCommand[expected];
-                }
-
-                if (message.OriginPlayerId != key.Origin || command.OperationId != key.Operation)
-                    continue;
-                if (command.CourseCount != expected)
-                {
-                    Mod.log.Warn("[MP] NetSync: dropping inconsistent course count for op=" +
-                                 key.Operation + " from player " + key.Origin + ".");
-                    continue;
-                }
-
-                int index = command.CourseIndex;
-                if (courses[index] != null) continue;
-                courses[index] = message;
-                decodedCourses[index] = command;
-                received++;
-            }
-
-            if (expected == 0) return false;
-
-            if (received != expected)
-            {
-                long deadline;
-                if (!_operationAssemblyDeadlines.TryGetValue(key, out deadline))
-                {
-                    deadline = now + OperationAssemblyWindowMs;
-                    _operationAssemblyDeadlines[key] = deadline;
-                }
-
-                if (now < deadline)
-                {
-                    RequeueAtFront(scanned);
-                    return false;
-                }
-
-                _operationAssemblyDeadlines.Remove(key);
-                var later = new List<SimulationCommandMessage>();
-                for (int i = 0; i < scanned.Count; i++)
-                {
-                    if (scanned[i].CommandId != NetPlacementCommand.Id)
-                    {
-                        later.Add(scanned[i]);
-                        continue;
-                    }
-                    NetPlacementCommand command;
-                    try { command = NetPlacementCommand.Decode(scanned[i].Body); }
-                    catch { continue; }
-                    if (scanned[i].OriginPlayerId != key.Origin || command.OperationId != key.Operation)
-                        later.Add(scanned[i]);
-                }
-                RequeueAtFront(later);
-                Diagnostics.FlightRecorder.Note("net incomplete op dropped=" + key.Operation +
-                    " courses=" + received + "/" + expected);
-                SyncInbox.RequestResync(Diagnostics.ResyncReport
-                    .Create("incomplete net operation expired", "net",
-                        Diagnostics.ResyncEvidence.StreamLoss)
-                    .About("op " + key.Operation + " from player " + key.Origin)
-                    .Tried("waited " + (OperationAssemblyWindowMs / 1000) +
-                           " s for the missing pieces of the road the other player drew")
-                    .Fact("pieces received", received + " of " + expected));
-                return false;
-            }
-
-            _operationAssemblyDeadlines.Remove(key);
-            operation = new List<SimulationCommandMessage>(expected);
-            nativeOperation = true;
-            bool hasNativeCourse = false;
-            bool hasGeometryOnlyCourse = false;
-            for (int i = 0; i < expected; i++)
-            {
-                operation.Add(courses[i]);
-                nativeOperation &= decodedCourses[i].HasNativeCourse;
-                hasNativeCourse |= decodedCourses[i].HasNativeCourse;
-                hasGeometryOnlyCourse |= !decodedCourses[i].HasNativeCourse;
-            }
-
-            // Preserve later operations in their original receive order. Extra messages carrying
-            // the completed key are duplicates or inconsistent fragments and are discarded.
-            var deferred = new List<SimulationCommandMessage>();
-            for (int i = 0; i < scanned.Count; i++)
-            {
-                if (scanned[i].CommandId != NetPlacementCommand.Id)
-                {
-                    deferred.Add(scanned[i]);
-                    continue;
-                }
-                NetPlacementCommand command;
-                try { command = NetPlacementCommand.Decode(scanned[i].Body); }
-                catch { continue; }
-                if (scanned[i].OriginPlayerId == key.Origin && command.OperationId == key.Operation)
-                    continue;
-                deferred.Add(scanned[i]);
-            }
-            RequeueAtFront(deferred);
-
-            // Current senders only group exact native definitions. Geometry-only capture represents
-            // one final edge per command. Rejecting mixed or grouped fallback input prevents a peer
-            // from smuggling a partially native operation into per-course fallback realization.
-            if ((hasNativeCourse && hasGeometryOnlyCourse) || (expected > 1 && !nativeOperation))
-            {
-                Diagnostics.FlightRecorder.Note("net incompatible multi-course op dropped=" +
-                                                  key.Operation);
-                SyncInbox.RequestResync(Diagnostics.ResyncReport
-                    .Create("incompatible net operation rejected", "net",
-                        Diagnostics.ResyncEvidence.StreamLoss)
-                    .About("op " + key.Operation + " from player " + key.Origin)
-                    .Tried("nothing - the operation mixed two course encodings that cannot be " +
-                           "applied as one transaction")
-                    .Fact("courses in the operation", expected));
-                operation = null;
-                nativeOperation = false;
-                return false;
-            }
-            return true;
-        }
-
-        private bool TryTakeNextPlacementMessage(out SimulationCommandMessage message)
-        {
-            if (DeferForTerrain)
-            {
-                message = default(SimulationCommandMessage);
-                return false;
-            }
-            if (_remoteDeferred.Count > 0)
-            {
-                message = _remoteDeferred[0];
-                _remoteDeferred.RemoveAt(0);
-                return true;
-            }
-            return _incoming.TryDequeue(out message);
-        }
-
-        private void RequeueAtFront(List<SimulationCommandMessage> messages)
-        {
-            if (messages != null && messages.Count > 0)
-                _remoteDeferred.InsertRange(0, messages);
-        }
-
-        /// <summary>
-        /// Re-queue an operation that is waiting for a target it cannot see yet, WITHOUT parking
-        /// the rest of the pipeline behind it.
-        ///
-        /// The queue is strictly ordered, so the previous behaviour - putting it straight back at
-        /// the front - stopped every later operation, from every player, for the whole retry
-        /// window. That is worse than a delay: in the sessions this came from, the deferred
-        /// placement spent ten seconds in front of a queue while the world it was searching went on
-        /// changing, and then asked for a full world reload because what it was looking for was no
-        /// longer there.
-        ///
-        /// Causal order was only ever meaningful per sender, so only ANOTHER sender's work may
-        /// overtake. Everything the same sender queued behind this operation stays behind it.
-        /// </summary>
-        private void RequeueStalledOperation(List<SimulationCommandMessage> messages)
-        {
-            if (messages == null || messages.Count == 0) return;
-            int origin = messages[0].OriginPlayerId;
-
-            // Admit what has arrived so the reorder sees the whole ready set rather than whatever a
-            // previous scan happened to leave behind. Realization stays gated where it always was
-            // (see TryTakeNextPlacementMessage); only where the messages sit changes, and the same
-            // inbox cap bounds it.
-            SimulationCommandMessage admitted;
-            while (_remoteDeferred.Count < NetInboxCap && _incoming.TryDequeue(out admitted))
-                _remoteDeferred.Add(admitted);
-
-            int insertAt = 0;
-            while (insertAt < _remoteDeferred.Count &&
-                   _remoteDeferred[insertAt].OriginPlayerId != origin) insertAt++;
-            _remoteDeferred.InsertRange(insertAt, messages);
-        }
-
-        /// <summary>
-        /// A hold whose window lapsed this long ago no longer counts. An operation can leave the
-        /// pipeline by other doors - duplicate suppression, a malformed decode - and a forgotten
-        /// hold must never be able to stop bulldozing for the rest of a session.
-        /// </summary>
-        private const long StaleHoldGraceMs = 2000;
-
-        /// <summary>
-        /// True while at least one native operation is inside a live window waiting for a target
-        /// that has not arrived, pruning any hold whose window has fully lapsed.
-        ///
-        /// The feeders that can only ever REMOVE such a target - bulldoze, road replacement - stand
-        /// down while this is true. They used to run ahead of it every frame, which is how a
-        /// placement came to be rejected for a road this machine had just deleted out from under
-        /// it. Two windows is the most any operation gets, so the hold is seconds, not minutes.
-        /// </summary>
-        public bool HasStalledNativeOperation(long now)
-        {
-            if (_nativeOperationHolds.Count == 0) return false;
-            List<NetOperationKey> stale = null;
-            bool live = false;
-            foreach (KeyValuePair<NetOperationKey, NativeOperationHold> entry in _nativeOperationHolds)
-            {
-                if (now < entry.Value.DeadlineMs + StaleHoldGraceMs) { live = true; continue; }
-                (stale ?? (stale = new List<NetOperationKey>())).Add(entry.Key);
-            }
-            if (stale != null)
-                for (int i = 0; i < stale.Count; i++) _nativeOperationHolds.Remove(stale[i]);
-            return live;
-        }
-
-        /// <summary>
-        /// Whether the resolver may use its relaxed last-resort matches for this operation - the
-        /// merged-node-as-edge-split fallback. Unlocked once a full retry window has passed, and
-        /// never locked again for that operation (see <see cref="NativeOperationHold"/>).
-        /// </summary>
-        private bool RelaxedResolveAllowed(NetOperationKey key, long now)
-        {
-            NativeOperationHold hold;
-            if (!_nativeOperationHolds.TryGetValue(key, out hold)) return false;
-            return hold.Relaxed || now >= hold.DeadlineMs;
-        }
-
-        /// <summary>
-        /// Arm or advance the hold on an operation whose target is missing. Returns true while it
-        /// should keep waiting; false once the current window is up and a verdict is due.
-        /// </summary>
-        private bool HoldUnresolvedOperation(NetOperationKey key, long now, long operationId,
-            string detail, out int windows)
-        {
-            NativeOperationHold hold;
-            if (!_nativeOperationHolds.TryGetValue(key, out hold))
-            {
-                hold = new NativeOperationHold
-                {
-                    DeadlineMs = now + NativeTargetRetryWindowMs,
-                    Relaxed = false,
-                    Windows = 1,
-                };
-                _nativeOperationHolds[key] = hold;
-                Diagnostics.FlightRecorder.Note("net native target retry op=" + operationId +
-                                                  " " + detail);
-            }
-            windows = hold.Windows;
-            return now < hold.DeadlineMs;
-        }
-
-        /// <summary>
-        /// Grant one more window after the arbiter declined to settle the report. Deliberately
-        /// shorter than the first: this one runs against a world the arbiter has frozen, so it is a
-        /// far better test than the first window was, and the feeders it holds up are waiting on it.
-        /// </summary>
-        private void ExtendUnresolvedOperation(NetOperationKey key, long now)
-        {
-            NativeOperationHold hold;
-            _nativeOperationHolds.TryGetValue(key, out hold);
-            hold.DeadlineMs = now + NativeTargetRetryWindowMs / 2;
-            hold.Relaxed = true;
-            hold.Windows++;
-            _nativeOperationHolds[key] = hold;
-        }
-
-        /// <summary>How a report and a withdrawal name the same stalled operation.</summary>
-        internal const string UnresolvedNativeTargetReason = "native net target did not resolve";
-        internal const string UnresolvedMixedTargetReason = "mixed net operation target did not resolve";
-
-        private static string NativeOperationSubject(long operationId, int origin) =>
-            "op " + operationId + " from player " + origin;
-
-        private static string MixedOperationSubject(long operationId, int origin) =>
-            "mixed op " + operationId + " from player " + origin;
-
-        /// <summary>
-        /// Release a hold because the operation succeeded. Withdrawing the report is the point of
-        /// holding one: a world reload that was proposed and then did not have to happen is worth
-        /// exactly as much in the log as one that did.
-        /// </summary>
-        private void ClearOperationHold(NetOperationKey key, string reason, string subject,
-            string outcome)
-        {
-            if (!_nativeOperationHolds.Remove(key)) return;
-            Diagnostics.ResyncArbiter.Withdraw("net", reason, subject,
-                Mod.Service != null ? Mod.Service.NowMs : 0L, outcome);
-        }
-
-        /// <summary>
-        /// What this machine actually has where the source anchored its endpoint, and why each
-        /// candidate was refused.
-        ///
-        /// This is the fact the log never carried. "No road within reach", "a Medium Road is there
-        /// instead of the Small Road named", and "the road is there but six metres lower" are three
-        /// entirely different bugs, and all three used to print the same line before reloading the
-        /// world. Runs once, on the frame a reload is being considered, so the wider search it does
-        /// costs nothing in the normal case.
-        /// </summary>
-        private string DescribeLocalAnchorNeighbourhood(NetEndpointIntent intent,
-            ref NodePool nodes, ref EdgePool edges)
-        {
-            const float SearchXZ = 16f;
-            float3 anchor = new float3(intent.AnchorX, intent.AnchorY, intent.AnchorZ);
-            var report = new System.Text.StringBuilder();
-
-            float bestNodeXZ = float.MaxValue, bestNodeDy = 0f;
-            NetCellIndex.Enumerator nodeCandidates = nodes.Index.Near(anchor.xz, SearchXZ);
-            while (nodeCandidates.MoveNext())
-            {
-                int i = nodeCandidates.Current;
-                float xz = math.distance(nodes.Data[i].m_Position.xz, anchor.xz);
-                if (xz >= bestNodeXZ) continue;
-                bestNodeXZ = xz;
-                bestNodeDy = nodes.Data[i].m_Position.y - anchor.y;
-            }
-
-            float bestEdgeXZ = float.MaxValue, bestEdgeDy = 0f;
-            string bestEdgePrefab = null;
-            NetCellIndex.Enumerator edgeCandidates = edges.Index.Near(anchor.xz, SearchXZ);
-            while (edgeCandidates.MoveNext())
-            {
-                int i = edgeCandidates.Current;
-                float t;
-                float xz = MathUtils.Distance(edges.Curves[i].m_Bezier.xz, anchor.xz, out t);
-                if (xz >= bestEdgeXZ) continue;
-                Entity edge = edges.Entities[i];
-                if (!EntityManager.Exists(edge)) continue;
-                bestEdgeXZ = xz;
-                bestEdgeDy = MathUtils.Position(edges.Curves[i].m_Bezier, t).y - anchor.y;
-                bestEdgePrefab = EntityManager.HasComponent<global::Game.Prefabs.PrefabRef>(edge)
-                    ? PrefabNameOf(EntityManager
-                        .GetComponentData<global::Game.Prefabs.PrefabRef>(edge).m_Prefab)
-                    : "(no prefab)";
-            }
-
-            if (bestEdgeXZ == float.MaxValue)
-            {
-                report.Append("no road at all within ").Append(SearchXZ.ToString("F0")).Append(" m");
-            }
-            else
-            {
-                report.Append("nearest road is '").Append(bestEdgePrefab).Append("' ")
-                    .Append(bestEdgeXZ.ToString("F1")).Append(" m away, ")
-                    .Append(bestEdgeDy.ToString("F1")).Append(" m in height");
-            }
-
-            if (bestNodeXZ != float.MaxValue)
-                report.Append("; nearest junction ").Append(bestNodeXZ.ToString("F1"))
-                    .Append(" m away, ").Append(bestNodeDy.ToString("F1")).Append(" m in height");
-
-            // The resolver's own tolerances, so a reader can see at a glance whether the candidate
-            // above was refused on distance or on identity.
-            report.Append(" (must be within ").Append(NativeEdgeResolveXZ.ToString("F0"))
-                .Append(" m and ").Append(NativeTargetResolveY.ToString("F0"))
-                .Append(" m of height, same prefab, same layers, same owner)");
-            return report.ToString();
-        }
-
-        private static NativeTargetRetryKey NativeRetryKey(SimulationCommandMessage message,
-            NetPlacementCommand command)
-        {
-            return new NativeTargetRetryKey
-            {
-                Origin = message.OriginPlayerId,
-                Operation = command.OperationId,
-                Course = command.CourseIndex,
-            };
-        }
-
-        /// <summary>
-        /// True when every point of <paramref name="span"/> already lies on live same-prefab geometry
-        /// - five samples along the curve, each of which must sit on SOME existing edge of that prefab
-        /// (the span may map to several local sub-edges). Uses the tight SplitMatch tolerances so a
-        /// parallel road or a span rebuilt at another elevation is never wrongly treated as a
-        /// duplicate.
-        /// </summary>
-        private bool SpanAlreadyBuilt(Entity prefab, Bezier4x3 span, ref EdgePool edges)
-        {
-            for (int s = 0; s <= 4; s++)
-            {
-                float3 p = MathUtils.Position(span, s / 4f);
-                bool covered = false;
-                NetCellIndex.Enumerator candidates = edges.Index.Near(p.xz, SplitMatch.TolXZ);
-                while (candidates.MoveNext())
-                {
-                    int i = candidates.Current;
-                    Bezier4x3 bez = edges.Curves[i].m_Bezier;
-                    float t;
-                    if (MathUtils.Distance(bez.xz, p.xz, out t) > SplitMatch.TolXZ) continue;
-                    if (math.abs(MathUtils.Position(bez, t).y - p.y) > SplitMatch.TolY) continue;
-                    if (EntityManager.GetComponentData<global::Game.Prefabs.PrefabRef>(edges.Entities[i]).m_Prefab
-                        != prefab) continue;
-                    covered = true;
-                    break;
-                }
-                if (!covered) return false;
-            }
-            return true;
-        }
-
-        /// <summary>
-        /// Native multi-course operations use the game's live network search tree for idempotence.
-        /// Only edges near each of the five coverage samples are visited, so a large grid scales with
-        /// local network density rather than with every edge in the city.
-        /// </summary>
-        private static bool SpanAlreadyBuilt(Entity prefab, Bezier4x3 span,
-            ref LiveEdgeSearchSnapshot search)
-        {
-            for (int s = 0; s <= 4; s++)
-            {
-                float3 point = MathUtils.Position(span, s / 4f);
-                var iterator = new SpanCoverageIterator
-                {
-                    Bounds = new Bounds3(
-                        point - new float3(SplitMatch.TolXZ, SplitMatch.TolY, SplitMatch.TolXZ),
-                        point + new float3(SplitMatch.TolXZ, SplitMatch.TolY, SplitMatch.TolXZ)),
-                    Point = point,
-                    Prefab = prefab,
-                    Curves = search.Curves,
-                    Prefabs = search.Prefabs,
-                    Owners = search.Owners,
-                    Temps = search.Temps,
-                    Deleted = search.Deleted,
-                };
-                search.Tree.Iterate(ref iterator);
-                if (!iterator.Covered) return false;
-            }
-            return true;
-        }
-
-        /// <summary>
-        /// True when the course's interior (away from both endpoints, which
-        /// <see cref="ClassifyEndpoint"/> already resolved) comes within splitting range of any
-        /// existing edge — a transversal crossing or a lengthwise overlap. The game cuts every such
-        /// edge during Temp generation, so the course counts against the one-splitting-course-per-batch
-        /// rule even though neither endpoint classifies as a split. The fallback probe uses native
-        /// connection layers and physical widths; a conservative false positive only serializes work,
-        /// while a false negative could place two conflicting split courses in one commit.
-        /// </summary>
-        private bool BodyTouchesExistingEdge(Bezier4x3 course, NetPrefabInfo placedInfo,
-            ref EdgePool edges)
-        {
-            // The control hull contains the curve, so an expanded-AABB miss is an exact reject.
-            float3 lo = math.min(math.min(course.a, course.b), math.min(course.c, course.d))
-                - new float3(MaxEndpointSearch, VerticalSnapTol, MaxEndpointSearch);
-            float3 hi = math.max(math.max(course.a, course.b), math.max(course.c, course.d))
-                + new float3(MaxEndpointSearch, VerticalSnapTol, MaxEndpointSearch);
-
-            // Sample tightly enough (≈ EdgeSnapDistance apart, via the control-polygon length upper
-            // bound) that a perpendicular crossing cannot slip between two samples.
-            float approxLen = math.distance(course.a, course.b) + math.distance(course.b, course.c)
-                + math.distance(course.c, course.d);
-            int samples = math.clamp((int)(approxLen / EdgeSnapDistance), 8, 128);
-
-            NetCellIndex.Enumerator candidates = edges.Index.Overlapping(lo.xz, hi.xz);
-            while (candidates.MoveNext())
-            {
-                int i = candidates.Current;
-                Bezier4x3 bez = edges.Curves[i].m_Bezier;
-                float3 elo = math.min(math.min(bez.a, bez.b), math.min(bez.c, bez.d));
-                float3 ehi = math.max(math.max(bez.a, bez.b), math.max(bez.c, bez.d));
-                if (math.any(elo > hi) || math.any(ehi < lo)) continue;
-
-                Entity candidate = edges.Entities[i];
-                NetPrefabInfo targetInfo = default(NetPrefabInfo);
-                if (EntityManager.HasComponent<global::Game.Prefabs.PrefabRef>(candidate))
-                    targetInfo = NetInfoOf(EntityManager.GetComponentData<global::Game.Prefabs.PrefabRef>(candidate).m_Prefab);
-                if (!LayersCanConnect(placedInfo, targetInfo)) continue;
-                float touchDistance = math.max(EdgeSnapDistance,
-                    placedInfo.HalfWidth + EdgeHalfWidth(candidate, targetInfo.HalfWidth) +
-                    placedInfo.SnapDistance);
-
-                for (int s = 1; s < samples; s++)
-                {
-                    float3 p = MathUtils.Position(course, s / (float)samples);
-                    // Endpoint neighbourhoods belong to endpoint classification (reuse/split/merge).
-                    if (math.distance(p.xz, course.a.xz) < NodeSnapDistance) continue;
-                    if (math.distance(p.xz, course.d.xz) < NodeSnapDistance) continue;
-                    float t;
-                    if (MathUtils.Distance(bez.xz, p.xz, out t) >= touchDistance) continue;
-                    if (math.abs(MathUtils.Position(bez, t).y - p.y) > VerticalSnapTol) continue; // other level
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Resolve where one course endpoint connects, in priority order: an existing real node (reuse),
-        /// a building's utility sub-net node (utility nets only - a power/pipe connector stub), a
-        /// pending new node another course in this batch creates (merge), a pending batch edge it taps
-        /// mid-span (defer until real), an existing real edge - reusing an end node for taps inside its
-        /// end zone, splitting for interior taps - else free ground. Returns the snap entity (node to
-        /// reuse, or edge to split, or Entity.Null) and, via out params, the split parameter and the
-        /// <c>Kind*</c> classification.
-        /// </summary>
-        private Entity ClassifyEndpoint(float3 p, NetPrefabInfo placedInfo,
-            ref NodePool nodes, ref EdgePool edges, ref NodePool ownedNodes,
-            NativeList<float3> batchNewNodes, NativeList<Bezier4x3> batchEdges,
-            out float t, out int kind)
-        {
-            t = 0f;
-            Entity node = FindNodeAt(p, placedInfo, ref nodes);
-            if (node != Entity.Null) { kind = KindReuseNode; return node; }
-            // A power line / pipe endpoint lying on a building's connector stub connects to it —
-            // the sender drew it onto that stub, so the committed segment ends exactly there.
-            if ((placedInfo.ConnectLayers & UtilityConnectLayers) != Layer.None)
-            {
-                node = FindUtilityNodeAt(p, ref ownedNodes, placedInfo);
-                if (node != Entity.Null) { kind = KindReuseConnector; return node; }
-            }
-            // Coincides with a new node another course in this batch creates -> leave it as a fresh node
-            // (Entity.Null) and let GenerateNodesSystem merge the two by exact position.
-            if (NearAny(p, batchNewNodes, NodeSnapDistance)) { kind = KindMergeBatch; return Entity.Null; }
-            // Taps the middle of an edge this batch is still building -> can't split a not-yet-real edge;
-            // defer the whole course to the next cycle, where that edge is real and this becomes a split.
-            if (MidSpanOfAnyBatch(p, batchEdges)) { kind = KindDeferBatchEdge; return Entity.Null; }
-            Entity edge, endNode;
-            FindEdgeAt(p, placedInfo, ref edges, out edge, out t, out endNode);
-            // A tap inside an existing edge's end zone reuses that end's node (see FindEdgeAt).
-            if (endNode != Entity.Null) { kind = KindReuseNode; return endNode; }
-            if (edge != Entity.Null) { kind = KindSplit; return edge; }
-            kind = KindFree;
-            return Entity.Null;
-        }
-
-        /// <summary>
-        /// Classify against the source's absolute height first. If that finds open ground and the
-        /// placed network is a utility, retry at the height produced by applying the source endpoint
-        /// elevation to this machine's surface. A successful second lookup means the endpoint is the
-        /// same visible local pipe/cable connection despite terrain or water drift; retaining the
-        /// first result for every other case preserves bridge/tunnel level separation.
-        /// </summary>
-        private Entity ClassifyEndpointWithLocalSurface(Entity prefab, float3 sourcePoint,
-            float2 sourceElevation, NetPrefabInfo placedInfo,
-            ref NodePool nodes, ref EdgePool edges, ref NodePool ownedNodes,
-            NativeList<float3> batchNewNodes, NativeList<Bezier4x3> batchEdges,
-            ref TerrainHeightData heightData, ref WaterSurfaceData<SurfaceWater> waterData,
-            out float t, out int kind)
-        {
-            Entity result = ClassifyEndpoint(sourcePoint, placedInfo, ref nodes, ref edges,
-                ref ownedNodes, batchNewNodes, batchEdges, out t, out kind);
-            if (kind != KindFree) return result;
-
-            float3 projected;
-            // Utility layers only here: this path classifies an endpoint the source left on open
-            // ground, where a height reference moved for a road is exactly the bridge/tunnel level
-            // separation that must be preserved.
-            if (!TryProjectEndpointToLocalSurface(prefab, placedInfo, sourcePoint,
-                    sourceElevation, false, ref heightData, ref waterData, out projected))
-                return result;
-
-            float projectedT;
-            int projectedKind;
-            Entity projectedResult = ClassifyEndpoint(projected, placedInfo,
-                ref nodes, ref edges, ref ownedNodes, batchNewNodes, batchEdges,
-                out projectedT, out projectedKind);
-            if (projectedKind == KindFree) return result;
-
-            t = projectedT;
-            kind = projectedKind;
-            _rzLocalSurfaceMatches++;
-            return projectedResult;
-        }
-
-        /// <summary>
-        /// Mark the echo-suppression guard for a course being realized. The capture side
-        /// consumes the key of the committed edge's START (its <c>a</c> endpoint), but the
-        /// committed geometry can differ from the command: an endpoint that reuses a node
-        /// lands exactly ON that node - up to <see cref="NodeSnapDistance"/> from the
-        /// commanded point, past the guard's 0.5 m buckets - a split lands on the split
-        /// point, and the game may commit the edge with its endpoints swapped. So mark
-        /// every position the committed start can be: both raw endpoints plus each end's
-        /// resolved snap target. Stale extras simply age out (15 s TTL).
-        /// </summary>
-        private void MarkRealizeGuards(string prefabName, float3 a, float3 d,
-            Entity startSnap, int startKind, float startT,
-            Entity endSnap, int endKind, float endT, long now)
-        {
-            _guard.Mark(ReplicationGuard.Key(prefabName, a), now);
-            _guard.Mark(ReplicationGuard.Key(prefabName, d), now);
-            MarkResolvedEndpoint(prefabName, startSnap, startKind, startT, now);
-            MarkResolvedEndpoint(prefabName, endSnap, endKind, endT, now);
-        }
-
-        private void MarkResolvedEndpoint(string prefabName, Entity snap, int kind, float t, long now)
-        {
-            if (snap == Entity.Null || !EntityManager.Exists(snap)) return;
-            float3 position;
-            if ((kind == KindReuseNode || kind == KindReuseConnector) && EntityManager.HasComponent<Node>(snap))
-                position = EntityManager.GetComponentData<Node>(snap).m_Position;
-            else if (kind == KindSplit && EntityManager.HasComponent<Curve>(snap))
-                position = MathUtils.Position(EntityManager.GetComponentData<Curve>(snap).m_Bezier, t);
-            else return;
-            _guard.Mark(ReplicationGuard.Key(prefabName, position), now);
-        }
-
-        // Diagnostic tally by endpoint classification.
-        private void TallyEnd(int kind)
-        {
-            switch (kind)
-            {
-                case KindReuseNode: _rzSnapEnds++; break;
-                case KindReuseConnector: _rzSnapEnds++; break;
-                case KindMergeBatch: _rzMergeEnds++; break;
-                case KindSplit: _rzMidEnds++; break;
-                default: _rzFreeEnds++; break;
-            }
-        }
-
-        private void TallySurfaceCorrection(float start, float end)
-        {
-            if (start != 0f) { _rzSurfaceCorrections++; _rzSurfaceCorrectionMax = math.max(_rzSurfaceCorrectionMax, math.abs(start)); }
-            if (end != 0f) { _rzSurfaceCorrections++; _rzSurfaceCorrectionMax = math.max(_rzSurfaceCorrectionMax, math.abs(end)); }
-        }
-
-        /// <summary>
-        /// Read this frame's terrain and water surfaces once per realize cycle. The water dependency
-        /// completes here so the data is main-thread readable; between simulation steps the handle is
-        /// already complete.
-        /// </summary>
-        private void TakeSurfaceSnapshot(ref TerrainHeightData heightData,
-            ref WaterSurfaceData<SurfaceWater> waterData)
-        {
-            heightData = _terrainSystem.GetHeightData(waitForPending: true);
-            JobHandle waterDeps;
-            waterData = _waterSystem.GetSurfaceData(out waterDeps);
-            waterDeps.Complete();
-        }
-
-        /// <summary>
-        /// True when <paramref name="p"/> lies within <paramref name="tol"/> (XZ) of any point at a
-        /// matching height. The height gate mirrors the game's node merge, which is by position - a
-        /// batch containing both a ground road and a bridge above it must not classify the bridge's
-        /// endpoint as merging with the ground node.
-        /// </summary>
-        private static bool NearAny(float3 p, NativeList<float3> points, float tol)
-        {
-            float2 xz = p.xz;
-            float tolSq = tol * tol;
-            for (int i = 0; i < points.Length; i++)
-                if (math.distancesq(xz, points[i].xz) < tolSq
-                    && math.abs(points[i].y - p.y) <= VerticalSnapTol) return true;
-            return false;
-        }
-
-        /// <summary>
-        /// True when <paramref name="point"/> taps the MIDDLE (away from both ends) of any curve this
-        /// batch is creating - the same mid-span test as <see cref="FindEdgeAt"/>, against pending
-        /// batch edges rather than real ones, with the same height gate (a crossing on another level
-        /// is not a tap).
-        /// </summary>
-        private static bool MidSpanOfAnyBatch(float3 point, NativeList<Bezier4x3> curves)
-        {
-            float2 p = point.xz;
-            for (int i = 0; i < curves.Length; i++)
-            {
-                Bezier4x3 bez = curves[i];
-                float tt;
-                if (MathUtils.Distance(bez.xz, p, out tt) >= EdgeSnapDistance) continue;
-                float3 sp = MathUtils.Position(bez, tt);
-                if (math.abs(sp.y - point.y) > VerticalSnapTol) continue;
-                if (math.distance(sp.xz, bez.a.xz) < MinSplitOffset) continue;
-                if (math.distance(sp.xz, bez.d.xz) < MinSplitOffset) continue;
-                return true;
-            }
-            return false;
         }
     }
 }
