+++ CS2MultiplayerMod/Game/Sync/Systems/Routes/RouteSyncSystem/Realize.cs
@@ -12,6 +12,12 @@ using CS2MultiplayerMod.Game.Sync.Infrastructure;
 
 namespace CS2MultiplayerMod.Game.Sync.Systems
 {
+    // Realizing a peer's route command: create, update or delete a transport line by driving the
+    // game's own route tool with the waypoints the sender used.
+    //
+    // Resolving those waypoints against local stops is in RealizeConnections.cs, finding the route
+    // a command refers to in RealizeMatch.cs, and the commit that finishes a create or update in
+    // RealizeCommit.cs.
     public partial class RouteSyncSystem
     {
         // Horizontal identity stays tight, because distinct stops of one prefab are metres apart.
@@ -424,687 +430,5 @@ namespace CS2MultiplayerMod.Game.Sync.Systems
                         "' number " + command.RouteNumber + ".");
             return RealizeResult.Applied;
         }
-
-        /// <summary>
-        /// A line may carry waypoints that only shape its path, but it is meaningless - and a sign
-        /// of a truncated graph - if it serves no stop at all.
-        /// </summary>
-        private bool ValidateRouteContract(Entity routePrefab,
-            RouteWaypointIntent[] waypoints, string prefabName)
-        {
-            if (waypoints == null || waypoints.Length < 2 ||
-                waypoints.Length > RouteCreateCommand.MaxWaypoints)
-            {
-                SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                    .Create("invalid route topology", "route",
-                        CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.Contradiction)
-                    .About("line topology")
-                    .Tried("nothing - the described line does not form a valid route here"));
-                return false;
-            }
-
-            if (!EntityManager.HasComponent<TransportLineData>(routePrefab)) return true;
-            for (int i = 0; i < waypoints.Length; i++)
-                if (!string.IsNullOrEmpty(waypoints[i].StopPrefabName))
-                    return true;
-
-            SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                .Create("public transport route without any stop rejected", "route",
-                    CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.Contradiction)
-                .About("line with no stops")
-                .Tried("nothing - a public transport line with no stops cannot be created"));
-            Mod.log.Warn("[MP] RouteSync rejected public-transport line '" + prefabName +
-                         "' because none of its waypoints is connected to a stop.");
-            return false;
-        }
-
-        /// <summary>
-        /// Maps each waypoint's portable stop identity onto a live local stop. Also returns the
-        /// waypoint positions to submit: a connected waypoint takes its resolved stop's own
-        /// transform, which is what the route tool records for a locally drawn line.
-        /// </summary>
-        private bool TryResolveConnections(Entity routePrefab,
-            RouteWaypointIntent[] waypoints, out Entity[] result, out float3[] positions,
-            out string failure)
-        {
-            failure = null;
-            result = new Entity[waypoints.Length];
-            positions = new float3[waypoints.Length];
-            bool needsStops = false;
-            for (int i = 0; i < waypoints.Length; i++)
-            {
-                positions[i] = WaypointPosition(waypoints[i]);
-                needsStops |= !string.IsNullOrEmpty(waypoints[i].StopPrefabName);
-            }
-            if (!needsStops) return true;
-
-            TransportLineData lineData = default(TransportLineData);
-            bool hasLineData = EntityManager.HasComponent<TransportLineData>(routePrefab);
-            if (hasLineData)
-                lineData = EntityManager.GetComponentData<TransportLineData>(routePrefab);
-
-            NativeArray<Entity> stops = _transportStops.ToEntityArray(Allocator.Temp);
-            try
-            {
-                var stopNames = new string[stops.Length];
-                var stopServes = new bool[stops.Length];
-                var stopPositions = new float3[stops.Length];
-                for (int s = 0; s < stops.Length; s++)
-                {
-                    // Two prefabs can share a name, so identity is checked on each candidate's own
-                    // prefab rather than by resolving the name to a single entity.
-                    Entity candidatePrefab =
-                        EntityManager.GetComponentData<PrefabRef>(stops[s]).m_Prefab;
-                    stopNames[s] = PrefabNameOf(candidatePrefab);
-                    stopServes[s] = !hasLineData || StopServesLine(candidatePrefab, lineData);
-                    stopPositions[s] = EntityManager
-                        .GetComponentData<global::Game.Objects.Transform>(stops[s]).m_Position;
-                }
-
-                for (int i = 0; i < waypoints.Length; i++)
-                {
-                    RouteWaypointIntent wanted = waypoints[i];
-                    if (string.IsNullOrEmpty(wanted.StopPrefabName)) continue;
-
-                    Entity best = Entity.Null;
-                    float3 bestPosition = default(float3);
-                    float bestScore = 0f;
-                    bool bestOwnerMatch = false;
-                    int sameName = 0;
-                    int wrongPurpose = 0;
-                    float nearest = float.MaxValue;
-                    float3 wantedPosition = StopPosition(wanted);
-
-                    for (int s = 0; s < stops.Length; s++)
-                    {
-                        Entity candidate = stops[s];
-                        if (!string.Equals(stopNames[s], wanted.StopPrefabName,
-                                StringComparison.Ordinal))
-                            continue;
-                        sameName++;
-                        if (!stopServes[s])
-                        {
-                            wrongPurpose++;
-                            continue;
-                        }
-
-                        float3 candidatePosition = stopPositions[s];
-                        nearest = math.min(nearest,
-                            math.distance(candidatePosition, wantedPosition));
-                        if (!StopPositionsMatch(candidatePosition, wantedPosition)) continue;
-
-                        // The owner is the only thing separating identical platforms of one
-                        // station, so an owner-identified candidate outranks an anonymous one.
-                        bool ownerMatch = StopOwnerMatches(candidate, wanted);
-                        float score = math.distancesq(candidatePosition, wantedPosition);
-                        bool better = best == Entity.Null ||
-                                      (ownerMatch && !bestOwnerMatch) ||
-                                      (ownerMatch == bestOwnerMatch && score < bestScore);
-                        if (!better) continue;
-                        best = candidate;
-                        bestPosition = candidatePosition;
-                        bestScore = score;
-                        bestOwnerMatch = ownerMatch;
-                    }
-
-                    if (best == Entity.Null)
-                    {
-                        failure = "waypoint " + i + " found no live '" + wanted.StopPrefabName +
-                                  "' stop near " + Describe(wantedPosition) + " (" + sameName +
-                                  " with that name" +
-                                  (wrongPurpose != 0
-                                      ? ", " + wrongPurpose + " not serving this line"
-                                      : string.Empty) +
-                                  (nearest < float.MaxValue
-                                      ? ", nearest " + nearest.ToString("0.0") + " m)"
-                                      : ")");
-                        return false;
-                    }
-                    result[i] = best;
-                    positions[i] = bestPosition;
-                }
-                return true;
-            }
-            finally
-            {
-                stops.Dispose();
-            }
-        }
-
-        /// <summary>
-        /// A line whose own purpose the stop does not serve is a blocking validation error in the
-        /// game, so a captured line never used one.
-        /// </summary>
-        private bool StopServesLine(Entity stopPrefab, TransportLineData lineData)
-        {
-            if (!EntityManager.HasComponent<TransportStopData>(stopPrefab)) return false;
-            TransportStopData stopData =
-                EntityManager.GetComponentData<TransportStopData>(stopPrefab);
-            return stopData.m_TransportType == lineData.m_TransportType &&
-                   (!lineData.m_PassengerTransport || stopData.m_PassengerTransport) &&
-                   (!lineData.m_CargoTransport || stopData.m_CargoTransport);
-        }
-
-        private string PrefabNameOf(Entity prefab)
-        {
-            string name;
-            if (_prefabNames.TryGetValue(prefab, out name)) return name;
-            name = _prefabSystem.GetPrefabName(prefab) ?? string.Empty;
-            _prefabNames[prefab] = name;
-            return name;
-        }
-
-        private static string Describe(float3 position) =>
-            "(" + position.x.ToString("0") + "," + position.y.ToString("0") + "," +
-            position.z.ToString("0") + ")";
-
-        private bool StopOwnerMatches(Entity stop, RouteWaypointIntent wanted)
-        {
-            if (string.IsNullOrEmpty(wanted.OwnerPrefabName)) return true;
-
-            Entity topOwner;
-            if (!TryFindTopOwner(stop, out topOwner) || topOwner == Entity.Null ||
-                !EntityManager.HasComponent<PrefabRef>(topOwner) ||
-                !EntityManager.HasComponent<global::Game.Objects.Transform>(topOwner))
-                return false;
-            string ownerName = PrefabNameOf(
-                EntityManager.GetComponentData<PrefabRef>(topOwner).m_Prefab);
-            if (!string.Equals(ownerName, wanted.OwnerPrefabName,
-                    StringComparison.Ordinal))
-                return false;
-            float3 ownerPosition = EntityManager
-                .GetComponentData<global::Game.Objects.Transform>(topOwner).m_Position;
-            return OwnerPositionsMatch(ownerPosition, OwnerPosition(wanted));
-        }
-
-        private void AddWaypointDefinitions(Entity definition, Entity[] connections,
-            float3[] positions, Entity originalRoute, bool appendClosure)
-        {
-            Entity[] originals = MatchOriginalWaypoints(originalRoute, connections, positions);
-            DynamicBuffer<WaypointDefinition> buffer =
-                EntityManager.AddBuffer<WaypointDefinition>(definition);
-            for (int i = 0; i < positions.Length; i++)
-            {
-                buffer.Add(new WaypointDefinition
-                {
-                    m_Position = positions[i],
-                    m_Connection = connections[i],
-                    m_Original = originals[i],
-                });
-            }
-
-            // Route generation reads a repeated first position as "this loop closes", and compares
-            // it exactly - so the closing entry repeats the same value, never a recomputed one.
-            if (appendClosure)
-            {
-                buffer.Add(new WaypointDefinition
-                {
-                    m_Position = positions[0],
-                    m_Connection = connections[0],
-                    m_Original = Entity.Null,
-                });
-            }
-        }
-
-        private Entity[] MatchOriginalWaypoints(Entity route, Entity[] connections,
-            float3[] positions)
-        {
-            var result = new Entity[positions.Length];
-            if (route == Entity.Null || !EntityManager.HasBuffer<RouteWaypoint>(route))
-                return result;
-
-            DynamicBuffer<RouteWaypoint> original =
-                EntityManager.GetBuffer<RouteWaypoint>(route, isReadOnly: true);
-            var used = new bool[original.Length];
-            for (int i = 0; i < positions.Length; i++)
-            {
-                for (int j = 0; j < original.Length; j++)
-                {
-                    if (used[j]) continue;
-                    Entity waypoint = original[j].m_Waypoint;
-                    if (!EntityManager.HasComponent<Position>(waypoint) ||
-                        !WaypointPositionMatches(
-                            EntityManager.GetComponentData<Position>(waypoint).m_Position,
-                            positions[i], connections[i]))
-                        continue;
-
-                    Entity oldConnection = Entity.Null;
-                    if (EntityManager.HasComponent<Connected>(waypoint))
-                        oldConnection =
-                            EntityManager.GetComponentData<Connected>(waypoint).m_Connected;
-                    if (oldConnection != connections[i]) continue;
-                    used[j] = true;
-                    result[i] = waypoint;
-                    break;
-                }
-            }
-            return result;
-        }
-
-        private bool RouteGraphMatches(Entity route, Entity[] connections, float3[] positions)
-        {
-            if (!EntityManager.HasBuffer<RouteWaypoint>(route)) return false;
-            DynamicBuffer<RouteWaypoint> current =
-                EntityManager.GetBuffer<RouteWaypoint>(route, isReadOnly: true);
-            if (current.Length != positions.Length) return false;
-
-            for (int i = 0; i < positions.Length; i++)
-            {
-                Entity waypoint = current[i].m_Waypoint;
-                if (!EntityManager.Exists(waypoint) ||
-                    !EntityManager.HasComponent<Position>(waypoint) ||
-                    !WaypointPositionMatches(EntityManager
-                            .GetComponentData<Position>(waypoint).m_Position,
-                        positions[i], connections[i]))
-                    return false;
-                Entity connection = Entity.Null;
-                if (EntityManager.HasComponent<Connected>(waypoint))
-                    connection =
-                        EntityManager.GetComponentData<Connected>(waypoint).m_Connected;
-                if (connection != connections[i]) return false;
-            }
-            return true;
-        }
-
-        /// <summary>
-        /// A waypoint that only shapes the path is placed at exactly the submitted position; one
-        /// bound to a stop follows that stop and is compared with the stop tolerances.
-        /// </summary>
-        private static bool WaypointPositionMatches(float3 actual, float3 wanted,
-            Entity connection) =>
-            connection == Entity.Null
-                ? math.distancesq(actual, wanted) <= FreeWaypointMatchDistanceSq
-                : StopPositionsMatch(actual, wanted);
-
-        private Entity FindExistingCreate(Entity prefab, int routeNumber,
-            RouteWaypointIntent[] desired, out bool numberConflict)
-        {
-            numberConflict = false;
-            if (routeNumber <= 0) return Entity.Null;
-
-            NativeArray<Entity> routes = _liveRoutes.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < routes.Length; i++)
-                {
-                    Entity candidate = routes[i];
-                    if (EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab != prefab ||
-                        RouteNumberOf(candidate) != routeNumber)
-                        continue;
-
-                    RouteSnapshot snapshot;
-                    if (TryCaptureSnapshot(candidate, out snapshot) &&
-                        WaypointsMatchIntent(snapshot.Waypoints, desired))
-                        return candidate;
-                    numberConflict = true;
-                    return Entity.Null;
-                }
-                return Entity.Null;
-            }
-            finally
-            {
-                routes.Dispose();
-            }
-        }
-
-        private HashSet<Entity> CaptureShapeMatches(Entity prefab,
-            RouteWaypointIntent[] desired)
-        {
-            var result = new HashSet<Entity>();
-            NativeArray<Entity> routes = _liveRoutes.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < routes.Length; i++)
-                {
-                    Entity candidate = routes[i];
-                    if (EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab != prefab)
-                        continue;
-                    RouteSnapshot snapshot;
-                    if (TryCaptureSnapshot(candidate, out snapshot) &&
-                        WaypointsMatchIntent(snapshot.Waypoints, desired))
-                        result.Add(candidate);
-                }
-                return result;
-            }
-            finally
-            {
-                routes.Dispose();
-            }
-        }
-
-        private Entity FindRoute(Entity prefab, int routeNumber, float3 anchor,
-            float maxAnchorDistanceSq, out bool ambiguous)
-        {
-            ambiguous = false;
-            var exact = new List<Entity>();
-            var spatial = new List<Entity>();
-            NativeArray<Entity> routes = _liveRoutes.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < routes.Length; i++)
-                {
-                    Entity candidate = routes[i];
-                    if (EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab != prefab)
-                        continue;
-                    if (routeNumber > 0 && RouteNumberOf(candidate) == routeNumber)
-                        exact.Add(candidate);
-
-                    float3 first;
-                    if (TryGetFirstWaypoint(candidate, out first) &&
-                        math.distancesq(first, anchor) <= maxAnchorDistanceSq)
-                        spatial.Add(candidate);
-                }
-            }
-            finally
-            {
-                routes.Dispose();
-            }
-
-            if (exact.Count == 1) return exact[0];
-            if (exact.Count > 1)
-            {
-                Entity match = Entity.Null;
-                for (int i = 0; i < exact.Count; i++)
-                {
-                    float3 first;
-                    if (!TryGetFirstWaypoint(exact[i], out first) ||
-                        math.distancesq(first, anchor) > maxAnchorDistanceSq)
-                        continue;
-                    if (match != Entity.Null)
-                    {
-                        ambiguous = true;
-                        return Entity.Null;
-                    }
-                    match = exact[i];
-                }
-                if (match != Entity.Null) return match;
-                ambiguous = true;
-                return Entity.Null;
-            }
-
-            if (spatial.Count == 1) return spatial[0];
-            ambiguous = spatial.Count > 1;
-            return Entity.Null;
-        }
-
-        private bool TryGetFirstWaypoint(Entity route, out float3 position)
-        {
-            position = default(float3);
-            if (!EntityManager.HasBuffer<RouteWaypoint>(route)) return false;
-            DynamicBuffer<RouteWaypoint> waypoints =
-                EntityManager.GetBuffer<RouteWaypoint>(route, isReadOnly: true);
-            if (waypoints.Length == 0 ||
-                !EntityManager.HasComponent<Position>(waypoints[0].m_Waypoint))
-                return false;
-            position =
-                EntityManager.GetComponentData<Position>(waypoints[0].m_Waypoint).m_Position;
-            return true;
-        }
-
-        private bool TryApplyMetadata(Entity route, Entity prefab, int routeNumber, uint rgba,
-            HashSet<Entity> ignoredNumberConflicts = null)
-        {
-            if (!RouteNumberAvailable(route, prefab, routeNumber,
-                    ignoredNumberConflicts))
-                return false;
-            if (routeNumber > 0)
-            {
-                if (EntityManager.HasComponent<RouteNumber>(route))
-                    EntityManager.SetComponentData(route,
-                        new RouteNumber { m_Number = routeNumber });
-                else
-                    EntityManager.AddComponentData(route,
-                        new RouteNumber { m_Number = routeNumber });
-            }
-
-            UnityEngine.Color32 color = UnpackColor(rgba);
-            if (EntityManager.HasComponent<Color>(route))
-                EntityManager.SetComponentData(route, new Color { m_Color = color });
-            else
-                EntityManager.AddComponentData(route, new Color { m_Color = color });
-            if (!EntityManager.HasComponent<Updated>(route))
-                EntityManager.AddComponent<Updated>(route);
-            return true;
-        }
-
-        private bool RouteNumberAvailable(Entity route, Entity prefab, int routeNumber,
-            HashSet<Entity> ignoredRoutes = null)
-        {
-            if (routeNumber <= 0) return true;
-            NativeArray<Entity> routes = _liveRoutes.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < routes.Length; i++)
-                {
-                    Entity other = routes[i];
-                    if (other == route ||
-                        (ignoredRoutes != null && ignoredRoutes.Contains(other)) ||
-                        EntityManager.GetComponentData<PrefabRef>(other).m_Prefab != prefab)
-                        continue;
-                    if (RouteNumberOf(other) == routeNumber) return false;
-                }
-                return true;
-            }
-            finally
-            {
-                routes.Dispose();
-            }
-        }
-
-        private void FinalizeCreatedRoutes(long now)
-        {
-            var claimed = new HashSet<Entity>();
-            var ready = new Dictionary<PendingCreateMetadata, Entity>();
-            for (int i = _pendingCreateMetadata.Count - 1; i >= 0; i--)
-            {
-                PendingCreateMetadata pending = _pendingCreateMetadata[i];
-                if (!pending.GraphCommitted) continue;
-                bool ambiguous;
-                Entity route = FindMetadataTarget(pending, claimed, out ambiguous);
-                if (route != Entity.Null)
-                {
-                    claimed.Add(route);
-                    ready.Add(pending, route);
-                    continue;
-                }
-
-                if (now < pending.DeadlineMs) continue;
-                _pendingCreateMetadata.RemoveAt(i);
-                SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                    .Create(ambiguous ? "ambiguous created route" : "created route did not materialize",
-                        "route", ambiguous
-                            ? CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.Contradiction
-                            : CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.MissingTarget)
-                    .About("created line " + pending.PrefabName + " number " + pending.RouteNumber)
-                    .Tried(ambiguous
-                        ? "declined to guess between two equally good candidate lines"
-                        : "waited for the line the game builds from this definition, not counting " +
-                          "time route realization was held back")
-                    .Fact("line number", pending.RouteNumber));
-                Mod.log.Warn("[MP] RouteSync could not finalize created line '" +
-                             pending.PrefabName + "' number " + pending.RouteNumber +
-                             "; requested a fresh world sync.");
-            }
-
-            // The game's initializer may temporarily give several routes created in one batch the
-            // same free number. Treat every route finalized here as a coordinated assignment, while
-            // still rejecting conflicts with established routes outside this batch.
-            var readyRoutes = new HashSet<Entity>(ready.Values);
-            foreach (KeyValuePair<PendingCreateMetadata, Entity> pair in ready)
-            {
-                PendingCreateMetadata pending = pair.Key;
-                Entity route = pair.Value;
-                _mutatedRoutesThisFrame.Add(route);
-                if (!TryApplyMetadata(route, pending.Prefab,
-                        pending.RouteNumber, pending.Rgba, readyRoutes))
-                {
-                    SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                        .Create("route number conflict after creation", "route",
-                            CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.Contradiction)
-                        .About("line number after creation")
-                        .Tried("assigned every line finalized in this batch together before rejecting the conflict"));
-                    Mod.log.Warn("[MP] RouteSync could not assign number " +
-                                 pending.RouteNumber + " to '" + pending.PrefabName +
-                                 "'; requested a fresh world sync.");
-                }
-                else
-                {
-                    RouteSnapshot snapshot;
-                    if (TryCaptureSnapshot(route, out snapshot))
-                        _knownRoutes[route] = snapshot;
-                    Diagnostics.FlightRecorder.Note("route create finalized number=" +
-                                                      pending.RouteNumber + " stops=" +
-                                                      pending.Waypoints.Length);
-                    Mod.Verbose("[MP] RouteSync finalized line '" +
-                                pending.PrefabName + "' number " +
-                                pending.RouteNumber + ".");
-                }
-                _pendingCreateMetadata.Remove(pending);
-            }
-        }
-
-        private Entity FindMetadataTarget(PendingCreateMetadata pending,
-            HashSet<Entity> claimed, out bool ambiguous)
-        {
-            ambiguous = false;
-            Entity shape = Entity.Null;
-            int shapeCount = 0;
-            NativeArray<Entity> routes = _liveRoutes.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < routes.Length; i++)
-                {
-                    Entity candidate = routes[i];
-                    if (claimed.Contains(candidate) ||
-                        (pending.PreexistingShapeMatches != null &&
-                         pending.PreexistingShapeMatches.Contains(candidate)) ||
-                        EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab !=
-                        pending.Prefab)
-                        continue;
-                    RouteSnapshot snapshot;
-                    if (!TryCaptureSnapshot(candidate, out snapshot) ||
-                        !WaypointsMatchIntent(snapshot.Waypoints,
-                            pending.Waypoints))
-                        continue;
-                    shape = candidate;
-                    shapeCount++;
-                }
-            }
-            finally
-            {
-                routes.Dispose();
-            }
-            if (shapeCount == 1) return shape;
-            ambiguous = shapeCount > 1;
-            return Entity.Null;
-        }
-
-        private bool DeleteStillNeedsRecovery(RouteDeleteCommand command)
-        {
-            Entity prefab;
-            if (!_prefabIndex.TryResolve(command.PrefabName, out prefab)) return true;
-            bool ambiguous;
-            Entity route = FindRoute(prefab, command.RouteNumber,
-                new float3(command.WaypointX, command.WaypointY, command.WaypointZ),
-                RouteAnchorMatchDistanceSq, out ambiguous);
-            return ambiguous || route != Entity.Null;
-        }
-
-        private void CompleteCreateCommit(PendingCreateMetadata pending)
-        {
-            if (!_pendingCreateMetadata.Contains(pending)) return;
-            pending.GraphCommitted = true;
-            if (Mod.Service != null)
-                pending.DeadlineMs = Mod.Service.NowMs + RetryWindowMs;
-            Diagnostics.FlightRecorder.Note("route create graph committed; awaiting identity");
-        }
-
-        private void ReplayCreateAfterCommitLoss(PendingCreateMetadata pending)
-        {
-            if (!_pendingCreateMetadata.Remove(pending)) return;
-            QueueCommitReplay(new PendingRouteCommand
-            {
-                Create = pending.Source,
-                OriginPlayerId = pending.OriginPlayerId,
-                DeadlineMs = pending.DeadlineMs,
-            }, "create");
-        }
-
-        private void CompleteUpdateCommit(PendingUpdateCommit pending)
-        {
-            if (_pendingUpdateCommit != pending) return;
-            _pendingUpdateCommit = null;
-
-            RouteSnapshot snapshot;
-            if (EntityManager.Exists(pending.Route) &&
-                TryCaptureSnapshot(pending.Route, out snapshot))
-                _knownRoutes[pending.Route] = snapshot;
-            else
-                _knownRoutes[pending.Route] = pending.Desired;
-            Diagnostics.FlightRecorder.Note("route update graph committed");
-        }
-
-        private void ReplayUpdateAfterCommitLoss(PendingUpdateCommit pending)
-        {
-            if (_pendingUpdateCommit != pending) return;
-            _pendingUpdateCommit = null;
-            QueueCommitReplay(new PendingRouteCommand
-            {
-                Update = pending.Source,
-                OriginPlayerId = pending.OriginPlayerId,
-                DeadlineMs = pending.DeadlineMs,
-            }, "update");
-        }
-
-        private void QueueCommitReplay(PendingRouteCommand command, string operation)
-        {
-            MultiplayerService service = Mod.Service;
-            long now = service != null ? service.NowMs : 0;
-            // Gameplay not being ready means the world is already being replaced. The replacement
-            // supersedes this commit, so asking for another one is noise - and asking for a world
-            // reload BECAUSE a world reload is under way is how a session gets into a loop.
-            if (service == null || !service.GameplaySyncReady)
-            {
-                Mod.log.Warn("[MP] RouteSync " + operation +
-                             " commit was lost while the world was being replaced; the incoming " +
-                             "world supersedes it.");
-                return;
-            }
-            if (now >= command.DeadlineMs ||
-                _pendingCommands.Count >= MaxPendingCommands)
-            {
-                SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                    .Create("route commit could not be replayed", "route",
-                        CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.StreamLoss)
-                    .About("route " + operation + " commit")
-                    .Tried("nothing - the armed commit was wiped and its window had already closed")
-                    .Fact("route commands still queued", _pendingCommands.Count));
-                Mod.log.Warn("[MP] RouteSync " + operation +
-                             " commit was lost and could not be replayed safely.");
-                return;
-            }
-
-            command.NextAttemptMs = now;
-            command.RetryDelayMs = InitialRetryDelayMs;
-            _pendingCommands.Insert(0, command);
-            Diagnostics.FlightRecorder.Note("route " + operation +
-                                              " commit re-queued");
-        }
-
-        private void MarkCreateGuards(RouteCreateCommand command, long now)
-        {
-            float3 first = WaypointPosition(command.Waypoints[0]);
-            _guard.Mark(RouteKey("route", command.PrefabName,
-                command.RouteNumber, first), now);
-            _guard.Mark(RouteShapeKey("route", command.PrefabName,
-                command.Waypoints), now);
-        }
-
-        private static uint PackColor(byte r, byte g, byte b, byte a) =>
-            (uint)(r | (g << 8) | (b << 16) | (a << 24));
-
-        private static UnityEngine.Color32 UnpackColor(uint rgba) =>
-            new UnityEngine.Color32((byte)rgba, (byte)(rgba >> 8),
-                (byte)(rgba >> 16), (byte)(rgba >> 24));
     }
 }
